Fix string handling
I still do not fully grasp the relation between `UA_Variant`,
`UA_String` and the underlying `char *`, but I have to use
`UA_Variant_setScalarCopy` to avoid (sometime!) getting garbage.

I just wanted to wrap a `char *` string, not duplicating it (after
all, that memory is kept mmapped the whole time), but clearly the `Copy`
suffix of the above API refers to something other than that string.
Let's hope to not leak something!